frontend/i18n: lang=... query parameter, refactor setLocale initialization
