+++ src/packages/frontend/app/render.tsx
@@ -6,13 +6,51 @@
 import ReactDOM from "react-dom";
 import { createRoot } from "react-dom/client";
 
-import { Redux } from "@cocalc/frontend/app-framework";
+import {
+  redux,
+  Redux,
+  useEffect,
+  useTypedRedux,
+} from "@cocalc/frontend/app-framework";
+import {
+  getLocale,
+  Locale,
+  LOCALIZATIONS,
+  OTHER_SETTINGS_LOCALE_KEY,
+} from "@cocalc/frontend/i18n";
+import { QueryParams } from "@cocalc/frontend/misc/query-params";
 import { AppContext, useAppContextProvider } from "./context";
-import { Localize } from "./localize";
+import { Localize, useLocalizationCtx } from "./localize";
 
 // App uses the context provided by Redux (for the locale, etc.) and Localize.
 function App({ children }) {
   const appState = useAppContextProvider();
+  const { setLocale } = useLocalizationCtx();
+  const other_settings = useTypedRedux("account", "other_settings");
+
+  useEffect(() => {
+    const lang = QueryParams.get("lang");
+    if (lang != null) {
+      if (lang in LOCALIZATIONS) {
+        console.warn(
+          `URL query parameter 'lang=${lang}' – overriding user configuration.`,
+        );
+        redux
+          .getActions("account")
+          .set_other_settings(OTHER_SETTINGS_LOCALE_KEY, lang);
+        setLocale(lang);
+      } else {
+        console.warn(
+          `URL query parameter 'lang=${lang}' provided, but not a valid locale.`,
+          `Known values: ${Object.keys(LOCALIZATIONS)}`,
+        );
+      }
+    } else {
+      const i18n: Locale = getLocale(other_settings);
+      setLocale(i18n);
+    }
+  }, [other_settings]);
+
   return <AppContext.Provider value={appState}>{children}</AppContext.Provider>;
 }
